Move buildParentData to library file
That way the HgLogGraph component stays small, and it's easier to see
what to modify when we want to create a GitLogGraph component.
